requirements.txt
google-adk == 1.0.0
immutabledict == 4.2.1
sqlglot == 26.10.1
db-dtypes == 1.4.2
regex == 2024.11.6
tabulate == 0.9.0
google-cloud-aiplatform 
absl-py == 2.2.2
pydantic == 2.11.3
deprecated
